refactor: refactor tests Home.test.tsx
