+++ src/pages/index.tsx
@@ -1,4 +1,4 @@
-import { FC } from 'react';
+import React, { FC } from 'react';
 import Catalog from '../components/Catalog/Catalog';
 import Popup from '../components/Popup/Popup';
 import { PopupProvider } from '../context/PopupContext';
